feat: Make colorblind simulators a poweruser-only feature
Add `showColorblindSimulators` to the array of poweruser features.

Add strings for the `showColorblindSimulators` feature in the poweruser pop-up to `core.yaml`.

Add `_initHide` property to `section.js` to enable initializing any section `<div>` with a `.hide` class in a non-permanent manner.

Call the `initHide()` function with D3.js chaining to set the `_initHide` property of the to `true`, which means that `uiSectionColorblindModeOptions` section should initially be hidden.

Add custom on-toggle behavior for `showColorblindSimulators` feature such that the state of the poweruser feature checkbox associated with it adds or removes the `.hide` class from the section's `<div>`.