Code organization / fix issues with MapLibre 3d viewer (closes #1441)
Lots going on in here.. The reported issue was that clicking on the "Show 3d
Map" checkbox takes the user to Null Island.

The cause of this was that the code to work with maplibre was spread around a
bunch of different places. Checking the box kicked off a render before the
maplibre map had actually moved into position, which emitted mapmove, which moved
the main map to Null Island.

This commit reorganizes things and fixes a bunch of miscellaneous issues.

- Map3dSystem.js is now responsible for almost everything related to the 3d map.
- We're a lot more careful about how startup works and when event handlers are set up
- map3d_viewer.js now _only_ creates the container.  It does this at UI startup
  (This must happen before MapLibre starts)
- Add Map3dSystem methods for things like `.render()`, `.visible`, `.toggle()`
- Have Map3dSystem respond to changes in the urlhash (so if a user manually
  adds `map3d=true` to the hash, it should become visible.
- Give the background list "extras" section an update selection so it can update the
  Map3d checkbox if needed
- When d3-transitioning from visible->invisible, we don't need to first set style like
  `display: block; opacity: 1`. It can start that transition from wherever it is.
  (in some situations this made the div flash visible and then disappear immediately)