Allow togging rapid features in map data panel (closes #1434)
This also freshens up some of the text in the help section that talks about
the Rapid features.  The old text was really geared towards "AI Roads" and we
now offer a bunch of other feature layers for other types.

The new wording more refers to the Rapid Features as "suggestions", a term we
use frequently when talking about Rapid.