Attempt at getting our UISystem render code to be more reentrant
Currently we can only really call render once, and this will *append* the
various components to the container.  Ideally we'd like to get to a place
where the components can re-render at run time - this would unblock switching
locales on the fly, re-rendering the components with different strings.
We're not there yet.

Some things in this commit includes:
- Remove the code guard that allows render to only get called once
- d3 data bind a [0] the top level elements, so they won't get recreated on a rerender
- Use enter selections to control which renders are the first time
- Anywhere where we do a keybind.on, do a keybind.off first, so we don't end up with duplicate keybindings.

This starts to address some of the issues, but doesn't get very far.

We can test this capability in the Chrome dev tools by forcing another render
to happen after Rapid is already started:

> ui = context.systems.ui
> container = context.container()
> container.call(ui.render)