Add underscores for "scoped" translations for imagery, community strings
Previously, `tagging` translations were loaded separately and their strings
needed a "scope" on the beginning of the string path.

Now imagery and community are like this too.

Keeping them in separate scopes means that we can fetch the translations
independently of the main rapid build with `core` scope.

By convention we need to access the scoped strings by prepending an
underscore and scope name to the beginning of the string path.