Switch sidebar and inspector to class components, use flexbox
This modernizes the components to be written in a way that we can more easily
support rerendering pieces of the app.

Flexbox also simplifies the sidebar calculations a lot, and should
help a lot with mobile and localization (e.g. ltr-rtl)

I'm also trying out a few new naming conventions that might help
us work with our (admittedly very confusing and complex) UI code:
- capital names for components (like React).      `inspector`    -> `Inspector`
- dollar prefix for d3-selections (like JQuery)   `sidebar`      -> `$sidebar`
- double dollar signs for _enter_ selections      `sidebarEnter` -> `$$sidebar`