Correctly sync the 'drawing' and 'active' pseudo classes
The main change here is in AbstractLayer and the various Photo PixiLayerXX.

What should happen is that these pseudo classes are applied when rendering.
As a layer processes its features, it calls `syncFeatureClasses` to determine
if any of the features needs some special styling.  This function will dirty
the feature if any of these classes had changed. But 'active'/'drawing' was
mixed up.

I think what might have happened before was:
 Drag Behavior set `feature.active = true` to avoid emitting events
 But it didn't set this 'active' class on the layer (to make it stick)
 Then eventually `syncFeatureClasses` would run and un-set that 'active' class
 This probably messed up the dragging so somebody changed `syncFeatureClasses`
  to treat `active` differently.
 But then this made the streetside markers weird, so somebody added extra code
  to the photo layers to look for which photo was 'active'

This commit tries to sort all of that out, and also update some docs and variable naming.