Improve style dirty when waiting for a marker texture to be ready
Sometimes we need to wait for a texture to be ready, so we need to leave
`_styleDirty=true` in that situation.

The previous code would actually leave `_styleDirty=true` even in situations
where the was no icon or marker, the new code better checks if we are
really waiting or not.

This had the surprising side effect of, always leaving viewfield markers dirty
(like the mapillary view cones) beccause they have no icon.

I could not figure out why the view cone could still rerender with updated
angles or get restyled without being specificaly set to be dirty -
it was because of this unintended side effect.

This should also make viewcone rendering a little more efficient.