Properly unclass all old selected photo when selecting a new photo
The unclass code was only considering the current layer, but it should really
unclass all photo layers.  This was causing for example, if you switch between
a streetside photo to a mapillary photo, it would leave the selection halo set
on both markers.

This also unrelated, updates the documentation for the
`photosUsed` and `imageryUsed` functions.