Switch from using 'active' class to new 'selectphoto' class
The previous few commits were leading up to this.  We use these pseudo classes
for a lot of things, both styling but also some behaviors.  They weren't
documented very well.

Applying the 'active' class to decide whether a photo was selected was causing all
kinds of problems, becasue that class also makes the feature non-interactive.

This had the effect of, sometimes you could switch from one photo marker to another,
it would leave the first one styled as selected, but also make it unclickable.

Just making a new pseudo class for photo styling means this isn't a problem
now, and it's a lot easier to keep track of which photo is the selected one.

Note that the 'selected' class also applies, it's the thing that draws the
dashed line halo around it.  Another feature can get 'selected', for example
if the user clicks on some OSM feature, but as long as the photo is still showing,
we stil want to keep the 'photoselected' class applied.