Fix for Mapillary Street Signs not appearing (closes #1513)
The root cause of this was checking `lastv` in the `loadTiles()` function.
This is something I added in v2.3 to avoid doing a lot of work to load data if
the viewport hasn't moved since the last `loadTiles()` call.
In the case of Mapillary, we actually may load up to 3 things: photos, detections, and signs
so each of these caches needs their own lastv, and can't use a shared lastv

This also includes other changes to make things a bit nicer:
- Rename MapillaryFeatures -> MapillaryDetections (because features can mean anything)
  We will call these things "detections" more consistently (not "objects", "features", "points")
- Clean up all the streetlevel translation strings, they were a mess
- Move the detections and signs markers up to the "qa" render group,
  so they are drawn above the streetlevel viewcones and sequences
  (this looks nicer and makes them easier to select)
- Fix the tests
- Lots of variable renames and jsdoc comments