Make the Mapillary detections actually selectable (closes #1517)
This makes it so that clicking on the detection.. actually enters select mode.
This also adds some UI to show the detection details in the inspector/sidebar.
- A header that says what the thing is alongside an icon
- A detail section that shows some information (currently just First Seen / Last Seen)

I've also cleaned up a bunch of the translations strings for working with
these detected objects.  I didn't realize that we've had these strings in
Rapid (and probably iD) going back a long time, we just never used them.