Show placeholder icon for detection where we don't have an icon. (closes #1518)
This commit adds a few more square textures that can stand in as placeholders.
If a detection type doesn't have a corresponding icon in the spritesheet, we'll
show a question mark icon '?'

I also added the detection "type" in the sidebar so people can see what kind of sign it is.

I also simplified the translation strings a bit, because we don't need a bunch of
duplicate "Unknown" or "Type" strings.  In situations where it's a common word used
throughout Rapid, we can just have that word exist once under `inspector`.