This improves the photoviewer layout and resizing
The goal is to be able to have all our photoviewers (Mapillary, Pannellum, Kartaview,
and any future photo viewers that we may add) sharing more common design and UI.

These photo viewers should all respond to resizing (of both the map and the
sidebar), work correctly LTR/RTL, and support having a responsive footer across
the bottom that shows any options, and any attribution and metadata about the photo.

Various parts of this were working/broken before, but it all works consistently now.