Replace default ticker listener as before
The variables are a little different in v8.
This is still hacky, but I'm not sure another way around it.

The `TickerPlugin` code that adds the Ticker to `PIXI.Application`
really wants to do `ticker.add(this.render, this, UPDATE_PRIORITY.LOW);`
even if we were to replace it with a different Ticker.

Our issue is we need to call `render` ourselves at the time of our choosing.