Replaced the AtlasAllocator code
This switches from using `@rapideditor/pixi-texture-allocator`
"fork of the plugin" to just including the code we need in Rapid.

The old code worked ok for v7, but the v8 upgrade was causing texture
corruption, and it became easier to just massively simplify things
than to debug and fix whatever was going on with the plugin code.