Use higher resolution textures for minimap (re: #285)
Something I mentioned in #285 and wanted to do a while ago.
The textures being displayed in the minimap are really designed to be
displayed on a large map.

This commit just doubles the rendered dimensions of the minimap and then
scales it down to the size that we want to display it at.