First step towards improving Pixi async init
There are 3 "special" elements that have traditionally been created
and "owned" by the `MapSystem`:
*  `supersurface`    The parent `div` for temporary transforms between redraws
*  `surface`         The sibling `canvas` map drawing surface
*  `overlay`         The sibling `div` overlay, offsets the supersurface transform

The timing of when these get created has caused issues when initializing Pixi
because it needs that `canvas` surface to first exist.

We can actually create these things early and then append them to the DOM at
render time, so we don't necessarily need to wait for a first render to occur
before Pixi starts up.

So this commit creates the elements early and then appends them in `render`.

This also changes from storing these things as `d3-selections` to instead
storing the actual `HTMLDivElement` and `HTMLCanvasElement` nodes.
Various places that worked on d3-selections now use the nodes directly.
We are also extending the convention that we prefix d3-selections with '$'.

This commit also removes some convenience functions from `Context`
Code that needs the `surface` or `surfaceRect` can get it from the `MapSystem`.