Fix issue causing incomplete render after resizing map larger
Since enabling multiView rendering in fa863d4cb, (needed for the minimap)
Now Pixi is rendering to a separate canvas before copying the result to the
main surface canvas. This separate canvas also needs a resize!

This should probably happpen more automatically, but for now I can just
include the command to ensure that the rendering canvas matches the target
canvas size.