Add SpectorJS debugging tool for WebGL inspection
This is added for the Rapid dev build only.
Enabling it is manual, run these commands in a browser dev console:

```
ui = context.systems.ui
ui.spector.toggle()
```

You can toggle it on and off.
The Red circle will initiate a capture on the next frame.
The built-in UI for Spector is a bit of a mess.
We can make this nicer if we need to.