Move a lot of UiSystem's code into new components UiMapFooter, UiOvermap
This commit works towards the goal of having the whole UI be reactive..

Much of the UI being managed by UiSystem was only setup to render on enter,
not on update. We'll eventually support updateable components, but for now
this splits a lot of the UI into smaller components that can render on update,
and can be improved independent of each other.