Move uiCmd -> utilCmd
This goes along with the other "UI cleanup" work that I've been doing,
trying to bring better consistency to the UI code.

It has always bothered me a bit since the early days that "uiCmd"
doesn't really belong with the other UI code.

Someday we should convert the keybinding code to a KeyboardSystem
and move what this function does into there.