Slightly better way to display the nearby contributors list
This switches to `display: inline-block` and adds
`max-width: 110px` and `text-overflow: ellipsis` to guard
against long usernames making this content too wide.

Also make sure the user-links include their parent span so the
:last-of-type selector can work to add commas between the list items.
Before it didn't work beacuse of the "and {n} others".
(the {n} becomes the last-of-type, and the last username got a trailing comma)