Update validator status and filter status footer displays (re: #449)
Both of these are class components now that can rerender as needed.

The validator status widget didn't seem to be updating its count correctly,
I think this is now fixed.

The filter status had been commented out since the v2 rewrite.  This commit
adds bunch of changes in the FilterSystem to improve a lot of what it does.
This was commented out because we were not actually gathering any statistics
because so much about how the scene is drawn now has changed from the v1 days.

One big issue with the old `FilterSystem.getMatches` code was that there was a
requirement that parent relations must get matched before child ways.
Now, `getMatches` will recurse up if it detects that a parent hasn't done its
matching before the child. These functions all have some agressive caching
so it performs well.

It would not surprise me if this was causing some issues with landuse or boundary
multipolygons not filtering their child ways correctly.