Allow the UI and Map to respond to locale changes (re: #1598)
This means that the user can change `&locale=xx` in the url hash
(or locale can change some other way) and the UI and map will redraw.

Not all of the UI is able to react to this yet, so we'll leave #1598 open,
but the main components in the tree down through the sidebar, main map,
controls, attributions, footer can handle it.

This commit also cleans up a few rtl issues in the map footer.