Avoid matching more kinds of things if a lifecycle tag is present (closes #1580)
This was reported as "Razed railways remain visible if historical objects
are hidden".  The reason this was happening was because for a thing to be
hidden, _all_ of its matches must be hidden (not _any_ of its matches).

That meant that to hide the razed railways, a user would need to enable
both the "railways" and "past_future" filters.

To make this work more the way people expect, we need to consider the
`this._isPastFuture(tags)` check in more places - any kind of physical
thing that might be razed or demolished.  That means that razed railways
won't match as "railways" anymore, they match only as "past_future".