Remove extra container nesting and old 'todo'
The reason for this was:
A long time ago each Layer had a `this.container`.  We moved away from this
so that we could sort things in the scene better, and switched to using
"group containers" in PixiScene to manage the sorting. The Layers themselves
still manage data but now add it to the appropriate "group container" for rendering.

MapUI is a special layer that lives above everything else.  We still had some
code around that would expect MapUI Layer to still have a main `.container`,
then the code would look for a child that it really wanted - for example

```js
const mapUIContainer = this.scene.layers.get('map-ui').container;
const selectedContainer = mapUIContainer.getChildByName('selected');
```

Now code like this is rewritten to just get what it needs instead of expecting
there to be an extra container with children.