Universal dataset catalog
We've had the catalog of Esri datasets, but the Microsoft and Meta datasets
were separate from this.

As we start offering more sources of data (Microsoft, Meta, Overture, and more?)
it would be great to get all the datasets into a single catalog that our users
can search through.  That's what this commit does!

- Introduce `RapidDataset` class to contain all the common props that a dataset needs
- Show RapidDatasets in the menus and catalog
- Add a lot of methods for working with datasets into RapidSystem
- RapidSystem now prepares the catalog at startup - it will getAvailableDatasets
  from the 3 data-providing services (`mapwithai`,`esri`,`overture`)
- This moves a bunch of responsibility around - some things that were in the
  UI code are now in RapidSystem - some stuff that was duplicated isn't anymore.

Still todo:
- make sure all the dataset have snippet, itemUrl, thumbnail (only the Esri ones still do)
- make sure as datasets get added, they default to a new color (currently they're all magenta)