Make sure to trigger an immediate render when enabling a layer.
This was a common race condition on any layers that depend on a service.

Before, in the layer `enable` methods, we would call `startAsync()` but never
actually chain off that Promise to trigger a redraw.  Enabling the layer _would_
trigger `layerchange` and a redraw, but if the layer code was still waiting for
`startAsync()` to settle, the redraw wouldn't do anything.

The effect was that you could toggle on a layer like streetside or mapillary,
but you wouldn't see any markers appear until you did something else later to
trigger another redraw, like move the mouse to hover over something.

This change also exposed a few situations where `startAsync()` wasn't actually
returning a chainable `Promise` - these return `Promise.resolve()` now.