Make sure to account for extra param in click handlers
Previously if the user clicked the Accept button it would pass the command datum
into the event handler as the second param, instead of the 'nextMode'.

Event handlers in d3 are passed (event, datum, ...) so we need to account
for both params even though we don't use the bound datum in this case.