WIP on #1650
The goal here is to eliminate the half-pixel correction that we used before to
close the seams between imagery tile textures packed into the atlas. see 4e8f16b

While it does prevent the colors from bleeding into the neighboring texture,
it distorts the image a little bit - this is especially visible when zoomed in.

Instead, we'll do this:
- Just add 1px of padding to everything packed into the atlases
- For imagery tiles, duplicate the edge data into this 1px padding, so when the tiles
  are sampled, they blend with the padding, not the neighboring texture.

I tried it and it works ok - but my method of filling the padding is inefficient.
I'm just calling `texSubImage2D` a bunch of times with the full image.

Will followup this commit with something more efficient that only writes the
pixels we need.