Eliminite the half-pixel correction that distorts imagery textures (closes #1650)
Followup from 3bd47239
Now all packed textures will include 1px of padding, and we'll duplicate each 1px
edge of pixels into the padding, so we should never sample colors from a neighboring
texture, and the image should never be distorted because of the half pixel uv correction.

The previous commit did a bunch of extra `texSubImage2D` to accomplish this,
now I'm just copying the raw pixels that we need, so seems pretty quick.

To make this easier, I'm changing the AtlasAllocator to only work with ImageData now
that we need access to pixels.  This means that any code that was trying to pack
something else needs to convert it to an ImageData first.

This commit also doubles up the resolution of the SVG symbols, since some of
them are used for Mapillary Signs, and these can have a lot of text/detail.
The Mapillary signs look noticeably better now.
Probably all the other SVG symbols got a slight visual improvement too.