Show a message when the graphics system has a context loss (closes #1658)
This shows a message in the red status bar like:
"The graphics system has encountered a problem.  Rapid will try to continue with reduced quality."

This changes the GraphicsSystem `contextlost`/`contextrestored` events to a more generic
`statuschange` event, this is more what other parts of Rapid like the UiStatus bar need.