MapRoulette: Fix Mustache property replacement by loading task geometries
Even though the method name is called "task" what it actually loaded was the MapRoulette "challenge". Which means we did not have the actual task properties to replace the Mustache tags.

Now we also load the task but only add the features which hold the properties. In MR one task has a FeatureCollection attached which holds geometries that are displayed in the UI and properties. Usually this is only one feature (there are special cases). We now flatten the properties of all features and use this list to replace the Mustache tags.