MapRoulette: Hide Challenge Desc when an explicit challenge is selected
The challenge description is a high level description on the challenge which is great when one select a MR node without knowing the challenge.

However, when a explicit challenge ID is selected, we can assume the challenge is known beforehand so this hides it to reduce clutter on the page.