MapRoulette: Improve sidebar design
The sidebar showed all information in one gray box, which made it hard to distinguish the challenge description and task instruction; especially when the challenge authors used headlines in their text as well.

This changes the UI to show description and instruction in separate boxes and updates the loading and error messages to work with this updated design.