Maproulette fixes and improvements (#1689)
* MapRoulette: Add missing translation for error message

* Rename css class for QA tools

We use the same class for the MapRoulette Sidepanel and the term error was confusing.

* MapRoulette: Reduce line height for Sidebar title

* MapRoulette: Fix Mustache property replacement by loading task geometries

Even though the method name is called "task" what it actually loaded was the MapRoulette "challenge". Which means we did not have the actual task properties to replace the Mustache tags.

Now we also load the task but only add the features which hold the properties. In MR one task has a FeatureCollection attached which holds geometries that are displayed in the UI and properties. Usually this is only one feature (there are special cases). We now flatten the properties of all features and use this list to replace the Mustache tags.

* MapRoulette: Only remove Mustache tag if replaced

This way it is clear that something was there but was not replaced. I think this is easier to understand than showing nothing.

* MapRoulette: Fix the markdown parsing

Before, the markdown parsing was done on the initial data. However, this data can include Mustache tags which are replaced later which also can hold Markdown.

Now we parse the Markdown after replacing all Mustache tags.

* MapRoulette: Fix `osmIdentifier` links which interact with the map

The Mustache replacement also handels a special case of `osmIdentifier` which create a special link that hovers or selects the element in the map.

This change fixes this feature for all cases when the OSM identifier follows the pattern `way/${number}`, `node/${number}`, `relation/${number}`

* MapRoulette: Hide Challenge Desc when an explicit challenge is selected

The challenge description is a high level description on the challenge which is great when one select a MR node without knowing the challenge.

However, when a explicit challenge ID is selected, we can assume the challenge is known beforehand so this hides it to reduce clutter on the page.

* MapRoulette: Apply default styles to Markdown

The markdown that is provided by the challenge authors can add headlines and lists to the sidebar. This adds some basic styling for this case.

* MapRoulette: Improve sidebar design

The sidebar showed all information in one gray box, which made it hard to distinguish the challenge description and task instruction; especially when the challenge authors used headlines in their text as well.

This changes the UI to show description and instruction in separate boxes and updates the loading and error messages to work with this updated design.

* MapRoulette: Add error logging

This makes it easier to debug the error cases when the MR API returns an error.